feat(LT-3852): moving RFQ fetch SQL code into stored procedure

+++ src/MarginTrading.Backend.Contracts/Common/PaginatedResponseContract.cs
@@ -32,9 +32,9 @@ namespace MarginTrading.Backend.Contracts.Common
         /// <summary>
         /// Total size of all the contents
         /// </summary>
-        public int TotalSize { get; }
+        public long TotalSize { get; }
 
-        public PaginatedResponseContract([NotNull] IReadOnlyList<T> contents, int start, int size, int totalSize)
+        public PaginatedResponseContract([NotNull] IReadOnlyList<T> contents, int start, int size, long totalSize)
         {
             Contents = contents ?? throw new ArgumentNullException(nameof(contents));
             Start = start;
